+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
 export { sseEventsToMessages, sseEventsToMessagesAsync } from "./utils";
 export type * from "./types";
-
+export { ToolRegistry } from "./tool-registry";
 export { PROTOCOL_VERSION } from "./version";